tests/wpt/tests/scroll-animations/css/timeline-offset-keyframes-with-scroll-timeline.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<link rel="help" src="https://drafts.csswg.org/scroll-animations-1/#named-timeline-range">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/web-animations/testcommon.js"></script>
<script src="support/testcommon.js"></script>
<script src="/web-animations/resources/keyframe-utils.js"></script>
<title>Animation range and scroll timeline</title>
</head>
<style type="text/css">
  @keyframes anim {
    cover 100% {
      opacity: 1;
    }
    cover 0% {
      opacity: 0;
    }
    50% {
      opacity: 0.5;
    }
  }
  #scroller {
    border:  10px solid lightgray;
    overflow-y: scroll;
    overflow-x: hidden;
    width: 300px;
    height: 200px;
  }
  #target {
    margin-bottom: 800px;
    margin-top:  800px;
    margin-left:  10px;
    margin-right:  10px;
    width: 100px;
    height: 100px;
    z-index: -1;
    background-color: green;
    animation:  anim auto both linear;
    animation-timeline: scroll();
  }
</style>
<body>
  <div id="scroller">
    <div id="target"></div>
  </div>
</body>
<script>
  async function runTest() {
    promise_test(async t => {
      await waitForNextFrame();
      const anim = document.getAnimations()[0];
      await anim.ready;
      await waitForNextFrame();

      let frames = anim.effect.getKeyframes();
      let expected = [
        { offset: 0.5, computedOffset: 0.5, opacity: "0.5", easing: "linear",
          composite: "auto" },
        { offset: { rangeName: "cover", offset: CSS.percent(0) },
          computedOffset: 0, opacity:  "0", composite: "auto",
          easing: "linear" },
        { offset: { rangeName: "cover", offset: CSS.percent(100) },
          computedOffset: 1, opacity:  "1", composite: "auto",
          easing: "linear" }
      ];
      assert_frame_lists_equal(frames, expected);
    }, 'Keyframes with timeline-offsets reported using a scroll timeline');
  }

  window.onload = runTest;
</script>
